[#16] Implementing security for api-server
The API Server provides a security model that provides authentication
and authorization of incoming clients.

The security can be enabled/disabled (i.e. via `API_SERVER_SECURITY_ENABLED` env var)

Currently the api server support `jwt` token authentication.

The login api is defined in openapi.yml

/server/login

A client logs in to an api server by sending a POST request to the login path.
The request body contains login information (i.e. username and password for jwt authentication type)

Please refer to [api.md](api.md) for details of the login api.

Currently the security manager uses local file to store user's info.
The default users file name is `.users.json`
The users file name can be configured using `USERS_FILE_URL` env var.
See `.test.users.json` for sample values.

Currently the api server doesn't perform authorization on logged in users.

The server keeps a list of jolokia endpoints for clients to access.
The endpoints are loaded from a local file named
`.endpoints.json`. Each top level entry represents a jolokia endpoint.
An entry has a unique name and details to access the jolokia api.

See `.test.endpoints.json` for sample values.

When an authenticated client sends a request to the api-server, it
should present its token in the request header

    'Authorization: Bearer `token`'

It also need to give the `targetEndpoint` in the query part of the request
if the request is to access an jolokia endpoint.

For example `/execBrokerOperation?targetEndpoint=broker1`.

Direct Proxy means a client can pass a broker's endpoint info to the
api-server in order to access it via the api-server.

For example the [self-provisioning plugin]
(https://github.com/artemiscloud/activemq-artemis-self-provisioning-plugin)
uses this api to access the jolokia of a broker's jolokia endpoint.